examples: define DCFLAGS for AM65x and J7es
Some am65x & j7es Makefiles include DCFLAGS without defining it anywhere in the
file. Added a standardized comment explaining what DCFLAGS is in every Makefile
that uses the variable.

As of this commit, no changes were made to the labs/ Makefiles since none of the
labs Makefiles have DCFLAGS.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>