Add Release Notes
Each tag of the PRU Software Support Package (PSSP) comes with a description
of what changed between that tag and the previous tag. However, we have not had
a single Release Notes document that customers can reference to get a summary of
the entire history of the PSSP.

This commit adds a top level Release_Notes.txt file. It includes the information
from tags v5.1.0 - v5.8.0. This commit adds additional detail to the release notes
for tags v5.1.0 - v5.8.0 when the tag message was too vague to be useful.

Future tags will always be associated with a commit that specifically updates
the Release_Notes.txt file.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>